Fix incorrect cross axis layout size measuring when grid should have smaller size
When the number of items in the one line is smaller than max number of item
in the one line, the layout size must be sum of spacings and item sizes.

But, now the grid has size that bigger than expected.
The cause is recalculating of layout size.

This commit removes cross axis layout size recalculating.
The layout size is now fixed in the 'measure' method.