Convert OrientationIndependentConstraints to inline class
OrientationIndependentConstraints was a data class to store 4 size
values for measuring grid layout. It could be created tens of time in a
second when measuring. And it can cause GC pressure and laggy frame.

This commit change this data class to inline class with just long value.
Object allocation will be down to zero by inlining this value in measurement.